students/campus_feedback_app - Copy/requirements.txt
streamlit==1.38.0
pandas==2.2.2
pyyaml==6.0.2
vaderSentiment==3.3.2
textblob==0.19.0
openai>=1.40.0